requirements.txt
streamlit
torch
diffusers
pillow
numpy
matplotlib
transformers
torchvision